refactor location of celery log setup + write tests
