--- apps/common/__init__.py
@@ -1,4 +1,0 @@
-from apps.common.logger import setup_logs
-
-
-setup_logs()
